requirements.txt
pandas

py-cpuinfo

keras

tensorflow
